prepare: networksetup: testeth0 nmconnection backup

+++ prepare/envsetup/03_configure_networking.sh
@@ -140,7 +140,15 @@ configure_networking () {
             nmcli c modify testeth0 ipv4.route-metric 99 ipv6.route-metric 99
             sleep 1
             # Copy final connection to /tmp/testeth0 for later in test usage
-            yes 2>/dev/null | cp -rf /etc/sysconfig/network-scripts/ifcfg-testeth0 /tmp/testeth0
+            if ! test -f /tmp/nm_plugin_keyfiles; then
+                if [ ! -e /tmp/testeth0 ] ; then
+                    yes 2>/dev/null | cp -rf /etc/sysconfig/network-scripts/ifcfg-testeth0 /tmp/testeth0
+                fi
+            else
+                if ! test -f /tmp/testeth0; then
+                    yes 2>/dev/null | cp -rf /etc/NetworkManager/system-connections/testeth0.nmconnection /tmp/testeth0
+                fi
+            fi
 
         fi
